Extract basic SELECT tests to separate module
Refactored test organization for better maintainability:
- Created select_basic.rs with 3 basic SELECT tests
- Moved test_select_star, test_select_with_where, test_select_specific_columns
- Updated select_executor.rs to focus on advanced features (ORDER BY, JOIN, aggregates)
- Updated mod.rs with new module structure

All 21 tests passing.

Closes #26
