+++ crates/executor/src/tests/select_executor.rs
@@ -1,162 +1,9 @@
 //! SELECT executor tests
 //!
-//! Tests for SELECT statement execution including WHERE, JOIN, GROUP BY, and aggregates.
+//! Tests for SELECT statement execution including ORDER BY, JOIN, GROUP BY, and aggregates.
+//! Basic SELECT tests are in select_basic.rs.
 
 use super::super::*;
-#[test]
-fn test_select_star() {
-    let mut db = storage::Database::new();
-    let schema = catalog::TableSchema::new(
-        "users".to_string(),
-        vec![
-            catalog::ColumnSchema::new("id".to_string(), types::DataType::Integer, false),
-            catalog::ColumnSchema::new(
-                "name".to_string(),
-                types::DataType::Varchar { max_length: 100 },
-                true,
-            ),
-        ],
-    );
-    db.create_table(schema).unwrap();
-    db.insert_row(
-        "users",
-        storage::Row::new(vec![
-            types::SqlValue::Integer(1),
-            types::SqlValue::Varchar("Alice".to_string()),
-        ]),
-    )
-    .unwrap();
-    db.insert_row(
-        "users",
-        storage::Row::new(vec![
-            types::SqlValue::Integer(2),
-            types::SqlValue::Varchar("Bob".to_string()),
-        ]),
-    )
-    .unwrap();
-
-    let executor = SelectExecutor::new(&db);
-    let stmt = ast::SelectStmt {
-        select_list: vec![ast::SelectItem::Wildcard],
-        from: Some(ast::FromClause::Table { name: "users".to_string(), alias: None }),
-        where_clause: None,
-        group_by: None,
-        having: None,
-        order_by: None,
-        limit: None,
-        offset: None,
-    };
-
-    let result = executor.execute(&stmt).unwrap();
-    assert_eq!(result.len(), 2);
-    assert_eq!(result[0].values[0], types::SqlValue::Integer(1));
-    assert_eq!(result[0].values[1], types::SqlValue::Varchar("Alice".to_string()));
-    assert_eq!(result[1].values[0], types::SqlValue::Integer(2));
-    assert_eq!(result[1].values[1], types::SqlValue::Varchar("Bob".to_string()));
-}
-
-#[test]
-fn test_select_with_where() {
-    let mut db = storage::Database::new();
-    let schema = catalog::TableSchema::new(
-        "users".to_string(),
-        vec![
-            catalog::ColumnSchema::new("id".to_string(), types::DataType::Integer, false),
-            catalog::ColumnSchema::new("age".to_string(), types::DataType::Integer, false),
-        ],
-    );
-    db.create_table(schema).unwrap();
-    db.insert_row(
-        "users",
-        storage::Row::new(vec![types::SqlValue::Integer(1), types::SqlValue::Integer(25)]),
-    )
-    .unwrap();
-    db.insert_row(
-        "users",
-        storage::Row::new(vec![types::SqlValue::Integer(2), types::SqlValue::Integer(17)]),
-    )
-    .unwrap();
-    db.insert_row(
-        "users",
-        storage::Row::new(vec![types::SqlValue::Integer(3), types::SqlValue::Integer(30)]),
-    )
-    .unwrap();
-
-    let executor = SelectExecutor::new(&db);
-    let stmt = ast::SelectStmt {
-        select_list: vec![ast::SelectItem::Wildcard],
-        from: Some(ast::FromClause::Table { name: "users".to_string(), alias: None }),
-        where_clause: Some(ast::Expression::BinaryOp {
-            left: Box::new(ast::Expression::ColumnRef { table: None, column: "age".to_string() }),
-            op: ast::BinaryOperator::GreaterThanOrEqual,
-            right: Box::new(ast::Expression::Literal(types::SqlValue::Integer(18))),
-        }),
-        group_by: None,
-        having: None,
-        order_by: None,
-        limit: None,
-        offset: None,
-    };
-
-    let result = executor.execute(&stmt).unwrap();
-    assert_eq!(result.len(), 2);
-    assert_eq!(result[0].values[0], types::SqlValue::Integer(1));
-    assert_eq!(result[1].values[0], types::SqlValue::Integer(3));
-}
-
-#[test]
-fn test_select_specific_columns() {
-    let mut db = storage::Database::new();
-    let schema = catalog::TableSchema::new(
-        "users".to_string(),
-        vec![
-            catalog::ColumnSchema::new("id".to_string(), types::DataType::Integer, false),
-            catalog::ColumnSchema::new(
-                "name".to_string(),
-                types::DataType::Varchar { max_length: 100 },
-                true,
-            ),
-            catalog::ColumnSchema::new("age".to_string(), types::DataType::Integer, false),
-        ],
-    );
-    db.create_table(schema).unwrap();
-    db.insert_row(
-        "users",
-        storage::Row::new(vec![
-            types::SqlValue::Integer(1),
-            types::SqlValue::Varchar("Alice".to_string()),
-            types::SqlValue::Integer(25),
-        ]),
-    )
-    .unwrap();
-
-    let executor = SelectExecutor::new(&db);
-    let stmt = ast::SelectStmt {
-        select_list: vec![
-            ast::SelectItem::Expression {
-                expr: ast::Expression::ColumnRef { table: None, column: "name".to_string() },
-                alias: None,
-            },
-            ast::SelectItem::Expression {
-                expr: ast::Expression::ColumnRef { table: None, column: "age".to_string() },
-                alias: None,
-            },
-        ],
-        from: Some(ast::FromClause::Table { name: "users".to_string(), alias: None }),
-        where_clause: None,
-        group_by: None,
-        having: None,
-        order_by: None,
-        limit: None,
-        offset: None,
-    };
-
-    let result = executor.execute(&stmt).unwrap();
-    assert_eq!(result.len(), 1);
-    assert_eq!(result[0].values.len(), 2);
-    assert_eq!(result[0].values[0], types::SqlValue::Varchar("Alice".to_string()));
-    assert_eq!(result[0].values[1], types::SqlValue::Integer(25));
-}
 
 #[test]
 fn test_order_by_single_column_asc() {
